WIP: polish tensorflow submitter
