+++ python/sqlflow_submitter/tensorflow/estimator_example.py
@@ -84,7 +84,7 @@ if __name__ == "__main__":
         save="mymodel",
         batch_size=1,
         epochs=3,
-        verbose=2)
+        verbose=0)
     pred(is_keras_model=False,
         datasource=datasource,
         estimator=tf.estimator.DNNClassifier,
