+++ python/sqlflow_submitter/tensorflow/fast_predict.py
@@ -1,0 +1,46 @@
+# Copyright 2019 The SQLFlow Authors. All rights reserved.
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
+class FastPredict:
+    def __init__(self, estimator, input_fn):
+        self.estimator = estimator
+        self.first_run = True
+        self.closed = False
+        self.input_fn = input_fn
+
+    def _create_generator(self):
+        while not self.closed:
+            yield self.next_features[0], self.next_features[1]
+
+    def predict(self, feature_batch):
+        self.next_features = feature_batch
+        if self.first_run:
+            self.batch_size = len(feature_batch)
+            self.predictions = self.estimator.predict(
+                input_fn=self.input_fn(self._create_generator))
+            self.first_run = False
+        elif self.batch_size != len(feature_batch):
+            raise ValueError("All batches must be of the same size. First-batch:" + str(self.batch_size) + " This-batch:" + str(len(feature_batch)))
+
+        results = []
+        for _ in range(self.batch_size):
+            results.append(next(self.predictions))
+        return results
+
+    def close(self):
+        self.closed = True
+        try:
+            next(self.predictions)
+        except Exception as e:
+            print("Exception in fast_predict. This is probably OK: %s" % e)
